chore: Phuc - refactor code
