+++ src/Core/Modules/Product/Service/ProductEnum.cs
@@ -8,6 +8,6 @@
         public const string ADD_PRODUCT_SUCCESS = "Added product successfully";
         public const string UPDATE_PRODUCT_SUCCESS = "Updated product successfully";
         public const string DELETE_PRODUCT_SUCCESS = "Deleted product successfully";
-        public const string DELETE_PRODUCT_DENY = "Deleted product unsuccessfully";
+        public const string DELETE_PRODUCT_DENIED = "You are not allowed to delete product";
     }
 }
